Remove "Could not start watchman" warning
Historically, `watchman` was much more important with the use of in-project
`tmp` directory. With that approach, we had a lot more fsevent to filter
through.

Now, with the move to using the system's TMPDIR, this problem does not really
exist anymore.

On the other hand, since the user most likely did not have `watchman` already
installed on their computer, this "warning" is emitted to every new Ember user
when they start the server.

Since it sounded like something is broken, they would then have to spend time
understanding the issue (following the link) and installing the package, which
is quite involved on anything other than on OS X, all for (at this point, imo)
dubious benefits.

I propose that we remove this warning entirely. If we want to, in Ember CLI, we
could potentially count the fsevents or profile the filtering code, then emit
the warning only when we it becomes an issue. (Probably not worth it though?)

The trigger for this was https://github.com/ember-learn/guides-source/pull/1038,
which was caused by https://github.com/ember-learn/super-rentals-tutorial/pull/21.

Since we are now capturing the real output for the tutorial, it also captured
the warning (the omission previously was basically a "bug"). It felt like
_something_ need to be said about it in the tutorial text, since it looks like
something went wrong.

The options are:

1. Pre-install `watchman` on the build server, so the warning won't show up
2. Teach the readers to install `watchman`
3. Instruct the readers to ignore the warning

(1) feels like cheating without also doing (2).

(2) is tricky on anything other than OS X (i.e. Linux).

(3) is probably the best option, but that begs the question, why did we have
the warning in the first place then?

After a brief discussion on Discord, it seems like maybe the time for which
this warning is necessary has past, and even core team members routinely use
Ember without `watchman` installed, so removing it seemed appropriate.

* * *

Since the `watchmanSupportsPlatform` option is not used by anymore, I have
removed it. I would think that it would be used to skip certain checks, but
it is not. If that was a mistake I can bring it back.